+++ packages/backend/src/database/entities/emails.ts
@@ -8,11 +8,8 @@ export type DbEmail = {
     is_primary: boolean;
 };
 
-export type DbEmailIn = {
-    user_id: number;
-    email: string;
-    is_primary: boolean;
-};
+export type DbEmailIn = Pick<DbEmail, 'user_id' | 'email' | 'is_primary'>;
+export type DbEmailRemove = Pick<DbEmail, 'user_id' | 'email'>;
 
 // DB Errors:
 // user_id does not exist (foreign key)
@@ -20,4 +17,8 @@ export type DbEmailIn = {
 // user already has a primary address set (constraint)
 export const createEmail = async (db: Knex, emailIn: DbEmailIn) => {
     await db<DbEmail>('emails').insert<DbEmailIn>(emailIn);
+};
+
+export const deleteEmail = async (db: Knex, emailRemove: DbEmailRemove) => {
+    await db<DbEmail>('emails').where(emailRemove).delete();
 };