+++ packages/backend/src/database/entities/organizations.ts
@@ -1,4 +1,5 @@
 import { Knex } from 'knex';
+import database from '../database';
 
 export type DbOrganization = {
     organization_id: number;
@@ -7,9 +8,8 @@ export type DbOrganization = {
     created_at: Date;
 };
 
-export type DbOrganizationIn = {
-    organization_name: string;
-};
+export type DbOrganizationIn = Pick<DbOrganization, 'organization_name'>;
+export type DbOrganizationUpdate = Pick<DbOrganization, 'organization_name'>;
 
 // DB Errors: Unexpected response (no rows returned)
 export const createOrganization = async (
@@ -20,4 +20,13 @@ export const createOrganization = async (
         .insert<DbOrganizationIn>(organizationIn)
         .returning('*');
     return org[0];
+};
+
+export const updateOrganization = async (
+    organizationUuid: string,
+    organizationUpdate: DbOrganizationUpdate,
+): Promise<void> => {
+    await database<DbOrganization>('organizations')
+        .where('organization_uuid', organizationUuid)
+        .update<DbOrganizationUpdate>(organizationUpdate);
 };